code refactored to use a custom hook useDoubleClick
